+++ src/components/layouts/MyConcert/NewConcert.jsx
@@ -9,7 +9,7 @@ import Stack from "@mui/material/Stack";
 import { v4 as randomUUID } from "uuid";
 
 import { useSetRecoilState, useRecoilValue } from "recoil";
-import { concertsState, concertNamesState } from "@/pages/App";
+import { concertsState, concertNamesState } from "@/components/RecoilStates";
 
 const BoxButton = styled(ButtonBase)(({ theme }) => ({
   position: "relative",
