+++ src/components/layouts/MyConcertCardList.jsx
@@ -3,7 +3,7 @@ import Box from "@mui/material/Box";
 import MyConcertCard from "@/components/layouts/MyConcert/Card";
 import NewConcert from "@/components/layouts/MyConcert/NewConcert";
 
-import { concertListState } from "@/pages/App";
+import { concertListState } from "@/components/RecoilStates";
 import { useRecoilValue } from "recoil";
 
 export default function MyConcertCardList({ Data, clicknode, setClicknode }) {
